requirements.txt
flask
gunicorn
celery
Redis
Pillow
opencv-python-headless==4.5.5.62
moviepy
psycopg2
pydrive
python-dotenv